Refactor: move MemberValidator to domain service package
