+++ backend/src/main/java/com/tissue/api/member/application/service/command/MemberCommandService.java
@@ -8,6 +8,7 @@ import com.tissue.api.common.exception.type.DuplicateResourceException;
 import com.tissue.api.member.application.dto.SignupMemberCommand;
 import com.tissue.api.member.domain.model.Member;
 import com.tissue.api.member.domain.model.vo.Name;
+import com.tissue.api.member.domain.service.MemberValidator;
 import com.tissue.api.member.infrastructure.repository.MemberRepository;
 import com.tissue.api.member.presentation.dto.request.UpdateMemberEmailRequest;
 import com.tissue.api.member.presentation.dto.request.UpdateMemberPasswordRequest;
@@ -15,7 +16,6 @@ import com.tissue.api.member.presentation.dto.request.UpdateMemberProfileRequest
 import com.tissue.api.member.presentation.dto.request.UpdateMemberUsernameRequest;
 import com.tissue.api.member.presentation.dto.request.WithdrawMemberRequest;
 import com.tissue.api.member.presentation.dto.response.command.MemberResponse;
-import com.tissue.api.member.validator.MemberValidator;
 import com.tissue.api.security.PasswordEncoder;
 
 import lombok.RequiredArgsConstructor;
