+++ backend/src/main/java/com/tissue/api/member/domain/service/MemberValidator.java
@@ -1,4 +1,4 @@
-package com.tissue.api.member.validator;
+package com.tissue.api.member.domain.service;
 
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
