+++ backend/src/test/java/com/tissue/support/helper/ServiceIntegrationTestHelper.java
@@ -21,8 +21,8 @@ import com.tissue.api.issue.infrastructure.repository.IssueRepository;
 import com.tissue.api.issue.infrastructure.repository.IssueReviewerRepository;
 import com.tissue.api.member.application.service.command.MemberCommandService;
 import com.tissue.api.member.application.service.query.MemberQueryService;
+import com.tissue.api.member.domain.service.MemberValidator;
 import com.tissue.api.member.infrastructure.repository.MemberRepository;
-import com.tissue.api.member.validator.MemberValidator;
 import com.tissue.api.position.application.service.command.PositionCommandService;
 import com.tissue.api.position.application.service.command.PositionReader;
 import com.tissue.api.position.application.service.query.PositionQueryService;
